+++ apps/web-giddh/src/app/shared/header/components/account-add-new-details/account-add-new-details.component.html
@@ -738,6 +738,13 @@
                         </ng-container>
                     </ng-container>
                 </ng-container>
+                <a
+                    href="javascript:;"
+                    class="btn-link"
+                    (click)="goToCustomFields.emit(true)"
+                    *ngIf="!companyCustomFields?.length"
+                    >{{ commonLocaleData?.app_create_custom_field }}</a
+                >
             </tab>
             <div class="row">
                 <div class="col-sm-12 mt-5">
