+++ apps/web-giddh/src/app/shared/header/components/account-operations/account-operations.component.html
@@ -73,6 +73,8 @@
             [showBankDetail]="showBankDetail"
             [showVirtualAccount]="showVirtualAccount"
             [isDebtorCreditor]="isDebtorCreditor"
+            (goToCustomFields)="goToCustomFields.emit($event)"
+            [reloadCustomFields]="reloadCustomFields"
         >
         </account-add-new-details>
 
@@ -93,6 +95,8 @@
             [showBankDetail]="showBankDetail"
             [showVirtualAccount]="showVirtualAccount"
             [isDebtorCreditor]="isDebtorCreditor"
+            (goToCustomFields)="goToCustomFields.emit($event)"
+            [reloadCustomFields]="reloadCustomFields"
         >
         </account-update-new-details>
     </div>
