+++ apps/web-giddh/src/app/shared/header/components/account-update-new-details/account-update-new-details.component.html
@@ -759,6 +759,13 @@
                         </ng-container>
                     </ng-container>
                 </ng-container>
+                <a
+                    href="javascript:;"
+                    class="btn-link"
+                    (click)="goToCustomFields.emit(true)"
+                    *ngIf="!companyCustomFields?.length"
+                    >{{ commonLocaleData?.app_create_custom_field }}</a
+                >
             </tab>
         </tabset>
         <div class="row" *ngIf="!isOtherSelectedTab">
