+++ apps/web-giddh/src/app/shared/header/components/custom-fields/custom-fields.component.scss
@@ -1,0 +1,31 @@
+@import "../../../../../assets/css/color.scss";
+
+.custom-filed-list-wrapper {
+    padding: 15px;
+    height: 768px;
+    overflow-y: auto;
+    form {
+        max-width: 1280px;
+        width: 100%;
+    }
+    label {
+        line-height: 2.3;
+    }
+    .edit-close-btn {
+        .close-button {
+            margin-right: 15px;
+            i {
+                font-size: 18px;
+            }
+        }
+        i {
+            color: $black !important;
+        }
+    }
+}
+@media (max-height: 800px) {
+    .custom-filed-list-wrapper {
+        height: 550px;
+        overflow-y: auto;
+    }
+}
