+++ apps/web-giddh/src/app/shared/header/components/new-manage-groups-accounts/manage-groups-accounts.component.scss
@@ -3,29 +3,6 @@
 .modal-body .grp-col {
     max-width: 58.333%;
 }
-.custom-filed-list-wrapper {
-    padding: 15px;
-    height: 768px;
-    overflow-y: auto;
-    form {
-        max-width: 1280px;
-        width: 100%;
-    }
-    label {
-        line-height: 2.3;
-    }
-    .edit-close-btn {
-        .close-button {
-            margin-right: 15px;
-            i {
-                font-size: 18px;
-            }
-        }
-        i {
-            color: $black !important;
-        }
-    }
-}
 
 .height-in-vh {
     height: calc(100vh - 70px);
@@ -41,12 +18,7 @@
         color: $white !important;
     }
 }
-@media (max-height: 800px) {
-    .custom-filed-list-wrapper {
-        height: 550px;
-        overflow-y: auto;
-    }
-}
+
 @media (max-width: 767px) {
     .search-account {
         float: none !important;
