+++ apps/web-giddh/src/app/shared/header/components/new-manage-groups-accounts/manage-groups-accounts.component.ts
@@ -1,6 +1,6 @@
 import { debounceTime, takeUntil } from 'rxjs/operators';
 import { GroupsAccountSidebarComponent } from '../new-group-account-sidebar/groups-account-sidebar.component';
-import { AfterViewChecked, ChangeDetectorRef, Component, ElementRef, EventEmitter, HostListener, OnDestroy, OnInit, Output, Renderer2, TemplateRef, ViewChild } from '@angular/core';
+import { AfterViewChecked, ChangeDetectorRef, Component, ElementRef, EventEmitter, HostListener, OnDestroy, OnInit, Output, Renderer2, ViewChild } from '@angular/core';
 import { GroupsWithAccountsResponse } from '../../../../models/api-models/GroupsWithAccounts';
 import { AppState } from '../../../../store/roots';
 import { Store, select } from '@ngrx/store';
@@ -10,14 +10,7 @@ import { GroupWithAccountsAction } from '../../../../actions/groupwithaccounts.a
 import { GroupAccountSidebarVM } from '../new-group-account-sidebar/VM';
 import { PerfectScrollbarComponent } from 'ngx-perfect-scrollbar';
 import { GeneralService } from "../../../../services/general.service";
-import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
-import { IOption } from 'apps/web-giddh/src/app/theme/ng-select/ng-select';
-import { FormArray, FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { GroupService } from 'apps/web-giddh/src/app/services/group.service';
-import { ToasterService } from 'apps/web-giddh/src/app/services/toaster.service';
-import { BreakpointObserver } from '@angular/cdk/layout';
-import { cloneDeep } from 'apps/web-giddh/src/app/lodash-optimized';
-import { GeneralActions } from 'apps/web-giddh/src/app/actions/general/general.actions';
+import { TabsetComponent } from 'ngx-bootstrap/tabs';
 
 @Component({
     selector: 'app-manage-groups-accounts',
@@ -34,13 +27,12 @@ export class ManageGroupsAccountsComponent implements OnInit, OnDestroy, AfterVi
     @ViewChild('groupsidebar', { static: true }) public groupsidebar: GroupsAccountSidebarComponent;
     public config: PerfectScrollbarConfigInterface = { suppressScrollX: false, suppressScrollY: false };
     @ViewChild('perfectdirective', { static: true }) public directiveScroll: PerfectScrollbarComponent;
-
+    /** Tabset instance */
+    @ViewChild('staticTabs', { static: true }) public staticTabs: TabsetComponent;
     public breadcrumbPath: string[] = [];
     public breadcrumbUniquePath: string[] = [];
     public myModelRect: any;
     public searchLoad: Observable<boolean>;
-    /** model reference */
-    public modalRef: BsModalRef;
     public groupList$: Observable<GroupsWithAccountsResponse[]>;
     public currentColumns: GroupAccountSidebarVM;
     public psConfig: PerfectScrollbarConfigInterface;
@@ -50,43 +42,24 @@ export class ManageGroupsAccountsComponent implements OnInit, OnDestroy, AfterVi
     private destroyed$: ReplaySubject<boolean> = new ReplaySubject(1);
     /* This will hold if keyup for focus in search field is initialized */
     public keyupInitialized: boolean = false;
-    /* This will store if device is mobile or not */
-    public isMobileScreen: boolean = false;
-    /** Add custom field form reference */
-    public customFieldForm: FormGroup;
-    /** List custom row data type  */
-    public dataTypeList: IOption[] = [];
-    /** To check API call in progress */
-    public isGetCustomInProgress: boolean = true;
-    /** To check API call in progress */
-    public isSaveCustomInProgress: boolean = false;
-    /** To get any custom field in edit mode index */
-    public isEnabledIndex: number = null;
-    /** To get  custom fields length */
-    public updateModeLength: number = 0;
-    /** Index to delete row in custom field */
-    public selectedRowIndex: number = null;
-    /* To check form value validation */
-    public isCustomFormValid: boolean = true;
     /* This will hold local JSON data */
     public localeData: any = {};
     /* This will hold common JSON data */
     public commonLocaleData: any = {};
     /** True if initial component load */
     public initialLoad: boolean = true;
+    /** This holds active tab */
+    public activeTab: string = "master";
+    /** True if custom fields api needs to be called again */
+    public reloadCustomFields: boolean = false;
 
     // tslint:disable-next-line:no-empty
     constructor(
         private store: Store<AppState>,
         private groupWithAccountsAction: GroupWithAccountsAction,
-        private formBuilder: FormBuilder,
         private cdRef: ChangeDetectorRef,
-        private breakPointObservar: BreakpointObserver,
         private renderer: Renderer2,
-        private _generalService: GeneralService,
-        private modalService: BsModalService,
-        private groupService: GroupService,
-        private toasterService: ToasterService
+        private _generalService: GeneralService
     ) {
         this.searchLoad = this.store.pipe(select(state => state.groupwithaccounts.isGroupWithAccountsLoading), takeUntil(this.destroyed$));
         this.groupAndAccountSearchString$ = this.store.pipe(select(s => s.groupwithaccounts.groupAndAccountSearchString), takeUntil(this.destroyed$));
@@ -115,26 +88,8 @@ export class ManageGroupsAccountsComponent implements OnInit, OnDestroy, AfterVi
         }
     }
 
-    /**
-     * To open confirmation model
-     *
-     * @param {TemplateRef<any>} template
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public openModal(template: TemplateRef<any>, index: number): void {
-        this.modalRef = this.modalService.show(template);
-        this.selectedRowIndex = index;
-    }
-
     // tslint:disable-next-line:no-empty
     public ngOnInit() {
-        this.breakPointObservar.observe([
-            '(max-width: 767px)'
-        ]).pipe(takeUntil(this.destroyed$)).subscribe(result => {
-            this.isMobileScreen = result.matches;
-        });
-        this.customFieldForm = this.createCustomFieldForm();
-        this.getCompanyCustomField();
         // search groups
         this.groupSearchTerms.pipe(
             debounceTime(700), takeUntil(this.destroyed$))
@@ -221,233 +176,29 @@ export class ManageGroupsAccountsComponent implements OnInit, OnDestroy, AfterVi
     }
 
     /**
-     * To submit custom field data
-     *
-     * @param {string} operationType To check operation type
-     * @param {*} type API call operation type
-     * @param {*} value
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public submitCustomFields(value: any, operationType?: string): void {
-        this.isSaveCustomInProgress = true;
-        this.groupService.createCompanyCustomField(value.customField).pipe(takeUntil(this.destroyed$)).subscribe(response => {
-            if (response) {
-                if (response.status === 'success') {
-                    this.customFieldForm.get('customField').reset();
-                    let customFieldResponse = response.body;
-                    this.updateModeLength = customFieldResponse.length;
-                    this.renderCustomField(customFieldResponse);
-                    if (operationType === 'create') {
-                        this.toasterService.successToast(this.localeData?.custom_field_created);
-                    } else if (operationType === 'delete') {
-                        this.toasterService.successToast(this.localeData?.custom_field_deleted);
-                    } else {
-                        this.toasterService.successToast(this.localeData?.custom_field_updated);
-                    }
-                } else {
-                    this.toasterService.errorToast(response.message);
-                    this.getCompanyCustomField();
-                }
-                this.isEnabledIndex = null;
-                this.isSaveCustomInProgress = false;
-                if (this.modalRef) {
-                    this.modalRef.hide()
-                }
-            }
-        });
-    }
-
-    /**
-     * API call to get custom field data
-     *
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public getCompanyCustomField(): void {
-        this.isGetCustomInProgress = true;
-        this.groupService.getCompanyCustomField().pipe(takeUntil(this.destroyed$)).subscribe(response => {
-            this.isEnabledIndex = null;
-            if (response) {
-                if (response.status === 'success') {
-                    this.renderCustomField(response.body);
-                    this.updateModeLength = response.body.length;
-                } else if (response.message) {
-                    this.toasterService.errorToast(response.message);
-                }
-            }
-            this.isGetCustomInProgress = false;
-
-        });
-    }
-
-    /**
-     * To render custom field data
-     *
-     * @param {*} response
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public renderCustomField(response: any): void {
-        let res: any[] = response;
-        this.customFieldForm = this.createCustomFieldForm();
-        const customRow = this.customFieldForm.get('customField') as FormArray;
-        if (res.length) {
-            res.map(item => {
-                item.isEditMode = true;
-                customRow.push(this.initNewCustomField(item));
-            });
-            this.removeCustomFieldRow(0, false);
-        }
-    }
-
-    /**
-     * To create and initialize custom field form
-     *
-     * @returns {FormGroup}
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public createCustomFieldForm(): FormGroup {
-        return this.formBuilder.group({
-            customField: this.formBuilder.array([
-                this.initNewCustomField(null)
-            ])
-        });
-    }
-
-    /**
-     * To initialize custom field form row
-     *
-     * @returns {FormGroup}
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public initNewCustomField(item: any): FormGroup {
-        let initCustomForm = this.formBuilder.group({
-            key: [null, Validators.compose([Validators.required])],
-            dataType: [null, Validators.compose([Validators.required])],
-            valueLength: [null, Validators.compose([Validators.required])],
-            isEditMode: [false],
-            uniqueName: [null],
-        });
-        if (item) {
-            initCustomForm?.patchValue(item);
-        }
-        return initCustomForm;
-    }
-
-
-    /**
-    * To add new custom field row
-    *
-    * @returns {*}
-    * @memberof ManageGroupsAccountsComponent
-    */
-    public addNewCustomFieldRow(): any {
-        const customRow = this.customFieldForm.get('customField') as FormArray;
-        if (this.customFieldForm.valid) {
-            customRow.push(this.initNewCustomField(null));
-        } else {
-            this.toasterService.warningToast(this.localeData?.fill_mandatory_fields);
-        }
-        return;
-    }
-
-    /**
-     * To remove custom field form row
-     *
-     * @param {boolean} isUpdate To check for API call
-     * @param {number} index index number
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public removeCustomFieldRow(index: number, isUpdate: boolean): void {
-        if (!isUpdate) {
-            const row = this.customFieldForm.get('customField') as FormArray;
-            if (row.length > 0) {
-                row.removeAt(index);
-            }
-        } else {
-            const row = cloneDeep(this.customFieldForm.get('customField') as FormArray);
-            if (row.length > 0) {
-                row.removeAt(index);
-            }
-            let requestObject = {
-                customField: row.value
-            }
-            this.submitCustomFields(requestObject, 'delete');
-        }
-    }
-
-    /**
-     * To edit custom field row
-     *
-     * @param {number} index
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public editCustomfield(index: number): void {
-        const row = this.customFieldForm.get('customField') as FormArray;
-        this.isEnabledIndex = index;
-        row.controls[index].get('isEditMode').setValue(false);
-    }
-
-    /**
-     * To add remoive validation according to custom field type
-     *
-     * @param {*} event
-     * @param {number} index
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public customFieldTypeSelected(event: any, index: number) {
-        const row = this.customFieldForm.get('customField') as FormArray;
-        if (event.value === 'BOOLEAN') {
-            row.controls[index].get('valueLength').clearValidators();
-        } else {
-            if (event.value === 'NUMERIC') {
-                row.controls[index].get('valueLength').setValidators([Validators.required, Validators.max(30)]);
-            } else if (event.value === 'STRING') {
-                row.controls[index].get('valueLength').setValidators([Validators.required, Validators.max(150)]);
-            }
-        }
-        row.controls[index].get('valueLength').setValue(null);
-    }
-
-    /**
-     * To show check custom field validation
-     *
-     * @param {string} type
-     * @param {number} index
-     * @memberof ManageGroupsAccountsComponent
-     */
-    public checkValidation(type: string, index: number): void {
-        const row = this.customFieldForm.get('customField') as FormArray;
-        this.isCustomFormValid = true;
-        if (type === 'name') {
-            if (row.controls[index] && row.controls[index].get('key') && row.controls[index].get('key').value && row.controls[index].get('key').value.length > 50) {
-                this.toasterService.errorToast(this.localeData?.name_length_validation);
-                this.isCustomFormValid = false;
-            }
-        } else {
-            if (row.controls[index].get('dataType').value === 'NUMERIC' && row.controls[index].get('valueLength').value > 30) {
-                this.toasterService.warningToast(this.localeData?.number_length_validation);
-                this.isCustomFormValid = false;
-
-            } else if (row.controls[index].get('dataType').value === 'STRING' && row.controls[index].get('valueLength').value > 150) {
-                this.toasterService.warningToast(this.localeData?.string_length_validation);
-                this.isCustomFormValid = false;
-            }
-        }
-
-    }
-
-    /**
-     * Callback for translation response complete
+     * This will show custom fields tab if clicked create custom field from add/update account
      *
      * @param {boolean} event
      * @memberof ManageGroupsAccountsComponent
      */
-    public translationComplete(event: boolean): void {
-        if (event) {
-            this.dataTypeList = [
-                { label: this.commonLocaleData?.app_datatype_list?.string, value: "STRING" },
-                { label: this.commonLocaleData?.app_datatype_list?.number, value: "NUMERIC" },
-                { label: this.commonLocaleData?.app_datatype_list?.boolean, value: "BOOLEAN" }
-            ];
+    public showCustomFieldsTab(event: boolean) {
+        if(event) {
+            this.staticTabs.tabs[1].active = true;
         }
+    }
+
+    /**
+     * Callback for tab change
+     *
+     * @param {string} tab
+     * @memberof ManageGroupsAccountsComponent
+     */
+    public onTabChange(tab: string): void {
+        if(tab === "master" && this.activeTab === "custom") {
+            this.reloadCustomFields = true;
+        } else {
+            this.reloadCustomFields = false;
+        }
+        this.activeTab = tab;
     }
 }
