+++ apps/web-giddh/src/assets/locale/accounts/en.json
@@ -56,21 +56,7 @@
     "share_with": "Share with",
     "select_permission": "-Select permission-",
     "share_group": "Share Group",
-    "name": "Name",
-    "type": "Type",
-    "length": "Length",
-    "add_new_row": "+ Add New Row",
-    "delete_custom_field_title": "Are you sure, you want to delete this?",
-    "delete_custom_field_alert": "Deleting this field from here, will remove it from all the accounts",
     "master": "Master",
-    "name_placeholder": "Enter custom field name",
-    "custom_field_created": "Custom field created successfully",
-    "custom_field_deleted": "Custom field deleted successfully",
-    "custom_field_updated": "Custom field updated successfully",
-    "fill_mandatory_fields": "Please fill all mandatory field",
-    "name_length_validation": "Name can not be greater than 50 characters",
-    "number_length_validation": "Length can not be greater than 30 for type number",
-    "string_length_validation": "Length can not be greater than 150 for type string",
     "default_discounts": "Default Discounts",
     "select_discounts": "Select discounts",
     "all_inherited_discounts": "All Inherited Discounts",
@@ -84,7 +70,6 @@
         "custom": "Custom"
     },
     "export_group_ledger": "Export Group Ledger",
-    "field_type_placeholder": "Type to filter",
     "bank_details_error_message": "All the bank fields are mandatory if you provide data for any of them",
     "item_no": "Item #[ITEM_NO]"
 }
