+++ apps/web-giddh/src/assets/locale/accounts/hi.json
@@ -56,21 +56,7 @@
     "share_with": "के साथ शेयर करें",
     "select_permission": "अनुमति का चयन करें",
     "share_group": "शेयर ग्रुप",
-    "name": "नाम",
-    "type": "प्रकार",
-    "length": "लंबाई",
-    "add_new_row": "नई पंक्ति जोड़ें",
-    "delete_custom_field_title": "क्या आप वाकई इसे हटाना चाहते हैं?",
-    "delete_custom_field_alert": "इस फ़ील्ड को यहां से हटाते हुए, इसे सभी खातों से हटा दिया जाएगा",
     "master": "मास्टर",
-    "name_placeholder": "कस्टम फ़ील्ड नाम दर्ज करें",
-    "custom_field_created": "कस्टम फ़ील्ड सफलतापूर्वक बनाया गया",
-    "custom_field_deleted": "कस्टम फ़ील्ड सफलतापूर्वक हटा दिया गया",
-    "custom_field_updated": "कस्टम फ़ील्ड सफलतापूर्वक अपडेट किया गया",
-    "fill_mandatory_fields": "कृपया सभी अनिवार्य फ़ील्ड भरें",
-    "name_length_validation": "नाम 50 से अधिक वर्णों का नहीं हो सकता",
-    "number_length_validation": "टाइप संख्या के लिए लंबाई 30 से अधिक नहीं हो सकती है",
-    "string_length_validation": "टाइप स्ट्रिंग के लिए लंबाई 150 से अधिक नहीं हो सकती है",
     "default_discounts": "डिफ़ॉल्ट छूट",
     "select_discounts": "छूट का चयन करें",
     "all_inherited_discounts": "सभी अंतर्निहित छूट",
@@ -84,7 +70,6 @@
         "custom": "कस्टम"
     },
     "export_group_ledger": "निर्यात समूह लेजर",
-    "field_type_placeholder": "फ़िल्टर करने के लिए टाइप करें",
     "bank_details_error_message": "यदि आप उनमें से किसी के लिए डेटा प्रदान करते हैं तो सभी बैंक फ़ील्ड अनिवार्य हैं",
     "item_no": "मद #[ITEM_NO]"
 }
