+++ apps/web-giddh/src/assets/locale/accounts/mr.json
@@ -56,21 +56,7 @@
     "share_with": "सह सामायिक करा",
     "select_permission": "परवानगी निवडा का",
     "share_group": "गट सामायिक करा",
-    "name": "नाव",
-    "type": "प्रकार",
-    "length": "लांबी",
-    "add_new_row": "नवीन पंक्ती जोडा",
-    "delete_custom_field_title": "आपली खात्री आहे की आपण हे हटवू इच्छिता?",
-    "delete_custom_field_alert": "येथून हे फील्ड हटविण्यामुळे ते सर्व खात्यांमधून काढले जाईल",
     "master": "मास्टर",
-    "name_placeholder": "सानुकूल फील्ड नाव प्रविष्ट करा",
-    "custom_field_created": "सानुकूल फील्ड यशस्वीरित्या तयार केले",
-    "custom_field_deleted": " सानुकूल फील्ड यशस्वीरित्या काढला ",
-    "custom_field_updated": "सानुकूल फील्ड यशस्वीरित्या अद्यतनित केले",
-    "fill_mandatory_fields": "कृपया सर्व अनिवार्य फील्ड भरा",
-    "name_length_validation": "नाव 50 वर्णांपेक्षा जास्त असू शकत नाही",
-    "number_length_validation": "प्रकार संख्येसाठी लांबी 30 पेक्षा जास्त असू शकत नाही",
-    "string_length_validation": "टाइप स्ट्रिंगसाठी लांबी 150 पेक्षा जास्त असू शकत नाही",
     "default_discounts": "डीफॉल्ट सूट",
     "select_discounts": "सवलत निवडा",
     "all_inherited_discounts": "सवलत निवडा",
@@ -84,7 +70,6 @@
         "custom": "सानुकूल"
     },
     "export_group_ledger": "ग्रुप लेजर निर्यात करा",
-    "field_type_placeholder": "फिल्टर करण्यासाठी टाइप करा",
     "bank_details_error_message": "आपण त्यापैकी कोणासाठी डेटा प्रदान केल्यास सर्व बँक फील्ड अनिवार्य आहेत",
     "item_no": "आयटम #[ITEM_NO]"
 }
