+++ apps/web-giddh/src/assets/locale/custom-fields/hi.json
@@ -1,0 +1,17 @@
+{
+    "delete_custom_field_title": "क्या आप वाकई इसे हटाना चाहते हैं?",
+    "delete_custom_field_alert": "इस फ़ील्ड को यहां से हटाते हुए, इसे सभी खातों से हटा दिया जाएगा",
+    "name_placeholder": "कस्टम फ़ील्ड नाम दर्ज करें",
+    "custom_field_created": "कस्टम फ़ील्ड सफलतापूर्वक बनाया गया",
+    "custom_field_deleted": "कस्टम फ़ील्ड सफलतापूर्वक हटा दिया गया",
+    "custom_field_updated": "कस्टम फ़ील्ड सफलतापूर्वक अपडेट किया गया",
+    "fill_mandatory_fields": "कृपया सभी अनिवार्य फ़ील्ड भरें",
+    "name_length_validation": "नाम 50 से अधिक वर्णों का नहीं हो सकता",
+    "number_length_validation": "टाइप संख्या के लिए लंबाई 30 से अधिक नहीं हो सकती है",
+    "string_length_validation": "टाइप स्ट्रिंग के लिए लंबाई 150 से अधिक नहीं हो सकती है",
+    "name": "नाम",
+    "type": "प्रकार",
+    "length": "लंबाई",
+    "add_new_row": "नई पंक्ति जोड़ें",
+    "field_type_placeholder": "फ़िल्टर करने के लिए टाइप करें"
+}
