+++ apps/web-giddh/src/assets/locale/hi.json
@@ -654,5 +654,6 @@
         "weekly": "साप्ताहिक",
         "monthly": "महीने के",
         "unlimited": "असीमित"
-    }
+    },
+    "app_create_custom_field": "कस्टम फ़ील्डस बनाएं"
 }
