+++ apps/web-giddh/src/assets/locale/mr.json
@@ -655,5 +655,6 @@
         "weekly": "साप्ताहिक",
         "monthly": "मासिक",
         "unlimited": "अमर्यादित"
-    }
+    },
+    "app_create_custom_field": "सानुकूल फील्डस तयार करा"
 }
